LibKeyboard: Don't compile CharacterMap:set_system_map in kernel mode
In preparation for using Userspace<T> in Syscall::SC_setkeymap_params
remove the usage of SC_setkeymap_params from when compiling in kernel
mode. In kernel model we would need to do a bunch of explicit FlatPtr
cats to in order to get it to compile, and it's unused anyway, so just
avoid the pain.